Set the reporting port to the error port by default
This is important for the get-javascript command, which outputs the
result to stdout. In order for this to be useful, it can be redirected
to a file, but if reporting information also goes to stdout, then it
ends up in the resulting JavaScript. Switching these messages to stderr
prevents that problem.